<Implemented storeLocation>
Implemented code to receive id, name and location data from the driver
side of the app. Then add that update to the locations database, so it
can be accessed by the manager side of the app. The servlet will now
also check to see if the id provided exists in the database and if not,
add a new id to the database and return it to the app which will use
that id from now on. Additionally the servlet will also update the name
of the id in the ids table, if the name provided by the app has been
changed in the app.

+Added id checks to storeLocation
+Added create and return id to storeLocation in case of no pre-existing
id.
+Added code to change an id’s name to storeLocation