<Implemented Store Keg>
Added code that implements the servlet-side of keg tracking. The
servlet now takes in the user id, the id of a scanned keg, the user’s
lat and the user’s Lon.

The servlet then uses this info to check if the user id and the keg id
exist in the ids and KegKeys tables respectively. Then if not already
on the table, it adds the keg and driver to the table KegDrivers which
tracks which driver has which kegs. It also adds the current time and
entered lat Lon as the pickup location and time to a new entry on the
KegHistory table.

If the keg is already on KegDrivers then it is removed, and the current
time and entered lat Lon are entered into an existing entry on
KegHistory as the dropoff time and location.

+Added storeKeg().
#Modified doGet to accept request=storekeg.