<Added getKegData function to servlet>
Added a getKegData function to a servlet that returns all of the stored
info on the latest entry of a specific key id. Returning the latest
entry from KegHistory if it is not currently being transported, or the
latest entry from locations if it is.

+Added getKegData(String data) Method
+Added case “getkegdata” to doGet() switch statement

#Added Extra output strings to storeKeg method, intended to provide
extra debug info to the app when it attempts to store a keg with the
servlet.